requirements.txt
requests
pyjwt
cryptography

